ci: Relax build checks (#168)
I failed to notice that it’s very difficult to use SemVer-based versioning that relies on the git commit history to determine versioning if the build output _must_ be checked into source. Since we don’t plan to keep the build output in the source tree going forwards, I’m relaxing these checks to make it possible to merge future changes.