ci: Switch to Netlify for deployments (#156)
In order to make the PR process a little easier (and hopefully save a little money on Heroku), this change switches over to Netlify for build, and deployments. It removes the Python server that was added to support Heroku, and also removes the build directory as this is no longer required.