ci: Release master builds automatically (#181)
This change updates the GitHub Actions workflow to use the same build script as Netlify (with slightly different flags), and updates that script so that it can perform releases, conditional on the TRY_RELEASE environment variable (or --release flag). The workflow is updated to set TRY_RELEASE to true on master builds.