docs: Add historical changes to allow for full automated changelog generation (#194)
With the introduction of automated version number and release note generation, it's necessary to augment the history with changes that were tagged before conforming to Conventional Commits. This does that by adding `history.yaml` which brings across changes from the manual changelog, and replaces that changelog with an auto-generated one which includes the later releases.