docs: Remove the old changelog and link gameplaycolor.com (#195)
Due to the Semantic Versioning / CI lifecycle, the generated changelog will always be out-of-date (it should always be derived from the repository). This change removes the changelog to avoid confusion and instead adds a link to the auto-generated one on gameplaycolor.com. It also includes a drive-by fix to add links to the jQuery and GameBoy-Online projects.